fix: ensure the lnd service is always running
The lnd service has run into errors such as network timeouts which resulted in
the service being shut down (gracefully). When that happens, it results in the
service remaining unavailable until an administrator notices and takes action.
This change ensures the service is always restarted, not just in the event of a
crash.